refactor: improve V4L2 utility functions and update worker header guards
- Changed header guards in v4l2_utils.h and worker.h for consistency.
- Moved V4L2 utility functions from a class to a namespace for better organization.
- Updated main.cpp to use the new namespace for utility functions.
- Added media_query.cpp and media_query.h for handling media file queries and operations.
- Integrated media_query functions into the Conductor class for file management.
- Enhanced recorder functionality with file rotation and space checking.
- Updated CMakeLists.txt to include JPEG library and new media query source files.
- Removed unnecessary includes and improved code organization across various files.
