+++ src/codecs/jetson/jetson_scaler.h
@@ -6,7 +6,6 @@
 
 #include "codecs/frame_processor.h"
 #include "common/thread_safe_queue.h"
-#include "common/v4l2_utils.h"
 #include "common/worker.h"
 
 #include <NvBufSurface.h>
