+++ src/common/CMakeLists.txt
@@ -4,6 +4,7 @@ find_package(JPEG REQUIRED)
 include_directories(${JPEG_INCLUDE_DIR})
 
 set(COMMON_FILES
+    ${PROJECT_SOURCE_DIR}/jpeg_util.cpp
     ${PROJECT_SOURCE_DIR}/v4l2_frame_buffer.cpp
     ${PROJECT_SOURCE_DIR}/utils.cpp
     ${PROJECT_SOURCE_DIR}/v4l2_utils.cpp
@@ -12,10 +13,6 @@ set(COMMON_FILES
 
 if(JETSON_PLATFORM)
     set(MULTIMEDIA_API_BASE /usr/src/jetson_multimedia_api)
-
-    list(APPEND COMMON_FILES
-        ${PROJECT_SOURCE_DIR}/nv_utils.cpp
-    )
 endif()
 
 add_library(${PROJECT_NAME} ${COMMON_FILES})
@@ -29,10 +26,6 @@ endif()
 # Use libyuv in libwebrtc.a
 target_link_libraries(${PROJECT_NAME}
     ${WEBRTC_LIBRARY}
-    avformat
-    avcodec
-    avutil
-    swscale
     uuid
     ${JPEG_LIBRARIES}
 )
