+++ src/common/v4l2_utils.cpp
@@ -9,19 +9,23 @@
 #include <sys/mman.h>
 #include <unistd.h>
 
-bool V4L2Util::IsSinglePlaneVideo(v4l2_capability *cap) {
+namespace v4l2_util {
+
+namespace {
+
+bool IsSinglePlaneVideo(v4l2_capability *cap) {
     return (cap->capabilities & (V4L2_CAP_VIDEO_CAPTURE | V4L2_CAP_VIDEO_OUTPUT) &&
             (cap->capabilities & V4L2_CAP_STREAMING)) ||
            (cap->capabilities & V4L2_CAP_VIDEO_M2M);
 }
 
-bool V4L2Util::IsMultiPlaneVideo(v4l2_capability *cap) {
+bool IsMultiPlaneVideo(v4l2_capability *cap) {
     return (cap->capabilities & (V4L2_CAP_VIDEO_CAPTURE_MPLANE | V4L2_CAP_VIDEO_OUTPUT_MPLANE) &&
             (cap->capabilities & V4L2_CAP_STREAMING)) ||
            (cap->capabilities & V4L2_CAP_VIDEO_M2M_MPLANE);
 }
 
-std::string V4L2Util::FourccToString(uint32_t fourcc) {
+std::string FourccToString(uint32_t fourcc) {
     int length = 4;
     std::string buf;
     buf.resize(length);
@@ -34,197 +38,7 @@ std::string V4L2Util::FourccToString(uint32_t fourcc) {
     return buf;
 }
 
-int V4L2Util::OpenDevice(const char *file) {
-    int fd = open(file, O_RDWR);
-    if (fd < 0) {
-        throw std::runtime_error(std::string("Failed to open v4l2 device: ") + file);
-    }
-    DEBUG_PRINT("Successfully opened file %s (fd: %d)", file, fd);
-    return fd;
-}
-
-void V4L2Util::CloseDevice(int fd) {
-    close(fd);
-    DEBUG_PRINT("fd(%d) is closed!", fd);
-}
-
-bool V4L2Util::QueryCapabilities(int fd, v4l2_capability *cap) {
-    if (ioctl(fd, VIDIOC_QUERYCAP, cap) < 0) {
-        ERROR_PRINT("fd(%d) query capabilities: %s", fd, strerror(errno));
-        return false;
-    }
-    return true;
-}
-
-bool V4L2Util::InitBuffer(int fd, V4L2BufferGroup *gbuffer, v4l2_buf_type type, v4l2_memory memory,
-                          bool has_dmafd) {
-    v4l2_capability cap = {};
-    if (!V4L2Util::QueryCapabilities(fd, &cap)) {
-        return false;
-    }
-
-    DEBUG_PRINT("fd(%d) driver '%s' on card '%s' in %s mode", fd, cap.driver, cap.card,
-                V4L2Util::IsSinglePlaneVideo(&cap)  ? "splane"
-                : V4L2Util::IsMultiPlaneVideo(&cap) ? "mplane"
-                                                    : "unknown");
-    gbuffer->fd = fd;
-    gbuffer->type = type;
-    gbuffer->memory = memory;
-    gbuffer->has_dmafd = has_dmafd;
-
-    return true;
-}
-
-bool V4L2Util::DequeueBuffer(int fd, v4l2_buffer *buffer) {
-    if (ioctl(fd, VIDIOC_DQBUF, buffer) < 0) {
-        ERROR_PRINT("fd(%d) dequeue buffer: %s", fd, strerror(errno));
-        return false;
-    }
-    return true;
-}
-
-bool V4L2Util::QueueBuffer(int fd, v4l2_buffer *buffer) {
-    if (ioctl(fd, VIDIOC_QBUF, buffer) < 0) {
-        ERROR_PRINT("fd(%d) queue buffer(%u): %s\n", fd, buffer->type, strerror(errno));
-        return false;
-    }
-    return true;
-}
-
-bool V4L2Util::QueueBuffers(int fd, V4L2BufferGroup *gbuffer) {
-    for (int i = 0; i < gbuffer->num_buffers; i++) {
-        v4l2_buffer *inner = &gbuffer->buffers[i].inner;
-        if (!V4L2Util::QueueBuffer(fd, inner)) {
-            return false;
-        }
-    }
-    return true;
-}
-
-std::unordered_set<std::string> V4L2Util::GetDeviceSupportedFormats(const char *file) {
-    int fd = V4L2Util::OpenDevice(file);
-    v4l2_fmtdesc fmtdesc = {0};
-    fmtdesc.type = V4L2_BUF_TYPE_VIDEO_CAPTURE;
-    std::unordered_set<std::string> formats;
-
-    while (ioctl(fd, VIDIOC_ENUM_FMT, &fmtdesc) == 0) {
-        auto pixel_format = V4L2Util::FourccToString(fmtdesc.pixelformat);
-        formats.insert(pixel_format);
-        fmtdesc.index++;
-    }
-    V4L2Util::CloseDevice(fd);
-
-    return formats;
-}
-
-bool V4L2Util::SubscribeEvent(int fd, uint32_t type) {
-    v4l2_event_subscription sub = {};
-    sub.type = type;
-    if (ioctl(fd, VIDIOC_SUBSCRIBE_EVENT, &sub) < 0) {
-        ERROR_PRINT("fd(%d) does not support VIDIOC_SUBSCRIBE_EVENT(%d)", fd, type);
-        return false;
-    }
-    return true;
-}
-
-bool V4L2Util::SetFps(int fd, v4l2_buf_type type, uint32_t fps) {
-    struct v4l2_streamparm streamparms = {};
-    streamparms.type = type;
-    streamparms.parm.output.timeperframe.numerator = 1;
-    streamparms.parm.output.timeperframe.denominator = fps;
-    if (ioctl(fd, VIDIOC_S_PARM, &streamparms) < 0) {
-        ERROR_PRINT("fd(%d) set fps(%d): %s", fd, fps, strerror(errno));
-        return false;
-    }
-    return true;
-}
-
-bool V4L2Util::SetFormat(int fd, V4L2BufferGroup *gbuffer, uint32_t width, uint32_t height,
-                         uint32_t &pixel_format) {
-    v4l2_format fmt = {};
-    fmt.type = gbuffer->type;
-    ioctl(fd, VIDIOC_G_FMT, &fmt);
-
-    DEBUG_PRINT("fd(%d) original formats: %s(%dx%d)", gbuffer->fd,
-                V4L2Util::FourccToString(fmt.fmt.pix_mp.pixelformat).c_str(), fmt.fmt.pix_mp.width,
-                fmt.fmt.pix_mp.height);
-
-    if (width > 0 && height > 0) {
-        fmt.fmt.pix_mp.width = width;
-        fmt.fmt.pix_mp.height = height;
-        fmt.fmt.pix_mp.pixelformat = pixel_format;
-    }
-
-    if (ioctl(fd, VIDIOC_S_FMT, &fmt) < 0) {
-        ERROR_PRINT("fd(%d) set format(%s) : %s", fd,
-                    V4L2Util::FourccToString(fmt.fmt.pix_mp.pixelformat).c_str(), strerror(errno));
-        return false;
-    }
-
-    DEBUG_PRINT("fd(%d) latest format: %s(%dx%d)", gbuffer->fd,
-                V4L2Util::FourccToString(fmt.fmt.pix_mp.pixelformat).c_str(), fmt.fmt.pix_mp.width,
-                fmt.fmt.pix_mp.height);
-    // use the  return format
-    pixel_format = fmt.fmt.pix_mp.pixelformat;
-    gbuffer->num_planes = fmt.fmt.pix_mp.num_planes;
-
-    if (fmt.fmt.pix_mp.width != width || fmt.fmt.pix_mp.height != height) {
-        ERROR_PRINT("fd(%d) input size (%dx%d) doesn't match driver's output size (%dx%d): %s", fd,
-                    width, height, fmt.fmt.pix_mp.width, fmt.fmt.pix_mp.height, strerror(EINVAL));
-        throw std::runtime_error("the frame size doesn't match");
-    }
-
-    return true;
-}
-
-bool V4L2Util::SetCtrl(int fd, uint32_t id, int32_t value) {
-    v4l2_control ctrls = {};
-    ctrls.id = id;
-    ctrls.value = value;
-    if (ioctl(fd, VIDIOC_S_CTRL, &ctrls) < 0) {
-        ERROR_PRINT("fd(%d) set ctrl(%d): %s", fd, id, strerror(errno));
-        return false;
-    }
-    return true;
-}
-
-bool V4L2Util::SetExtCtrl(int fd, uint32_t id, int32_t value) {
-    v4l2_ext_controls ctrls = {};
-    v4l2_ext_control ctrl = {};
-
-    /* set ctrls */
-    ctrls.ctrl_class = V4L2_CTRL_CLASS_CODEC;
-    ctrls.controls = &ctrl;
-    ctrls.count = 1;
-
-    /* set ctrl*/
-    ctrl.id = id;
-    ctrl.value = value;
-
-    if (ioctl(fd, VIDIOC_S_EXT_CTRLS, &ctrls) < 0) {
-        ERROR_PRINT("fd(%d) set ext ctrl(%d): %s", fd, id, strerror(errno));
-        return false;
-    }
-    return true;
-}
-
-bool V4L2Util::StreamOn(int fd, v4l2_buf_type type) {
-    if (ioctl(fd, VIDIOC_STREAMON, &type) < 0) {
-        ERROR_PRINT("fd(%d) turn on stream: %s", fd, strerror(errno));
-        return false;
-    }
-    return true;
-}
-
-bool V4L2Util::StreamOff(int fd, v4l2_buf_type type) {
-    if (ioctl(fd, VIDIOC_STREAMOFF, &type) < 0) {
-        ERROR_PRINT("fd(%d) turn off stream: %s", fd, strerror(errno));
-        return false;
-    }
-    return true;
-}
-
-void V4L2Util::UnMap(V4L2BufferGroup *gbuffer) {
+void UnMap(V4L2BufferGroup *gbuffer) {
     for (int i = 0; i < gbuffer->num_buffers; i++) {
         if (gbuffer->buffers[i].dmafd > 0) {
             DEBUG_PRINT("close (%d) dmafd", gbuffer->buffers[i].dmafd);
@@ -238,7 +52,7 @@ void V4L2Util::UnMap(V4L2BufferGroup *gbuffer) {
     }
 }
 
-bool V4L2Util::MMap(int fd, V4L2BufferGroup *gbuffer) {
+bool MMap(int fd, V4L2BufferGroup *gbuffer) {
     for (int i = 0; i < gbuffer->num_buffers; i++) {
         V4L2Buffer *buffer = &gbuffer->buffers[i];
         v4l2_buffer *inner = &buffer->inner;
@@ -291,7 +105,183 @@ bool V4L2Util::MMap(int fd, V4L2BufferGroup *gbuffer) {
     return true;
 }
 
-bool V4L2Util::AllocateBuffer(int fd, V4L2BufferGroup *gbuffer, int num_buffers) {
+} // namespace
+
+int OpenDevice(const char *file) {
+    int fd = open(file, O_RDWR);
+    if (fd < 0) {
+        throw std::runtime_error(std::string("Failed to open v4l2 device: ") + file);
+    }
+    DEBUG_PRINT("Successfully opened file %s (fd: %d)", file, fd);
+    return fd;
+}
+
+void CloseDevice(int fd) {
+    close(fd);
+    DEBUG_PRINT("fd(%d) is closed!", fd);
+}
+
+bool QueryCapabilities(int fd, v4l2_capability *cap) {
+    if (ioctl(fd, VIDIOC_QUERYCAP, cap) < 0) {
+        ERROR_PRINT("fd(%d) query capabilities: %s", fd, strerror(errno));
+        return false;
+    }
+    return true;
+}
+
+bool InitBuffer(int fd, V4L2BufferGroup *gbuffer, v4l2_buf_type type, v4l2_memory memory,
+                bool has_dmafd) {
+    v4l2_capability cap = {};
+    if (!QueryCapabilities(fd, &cap)) {
+        return false;
+    }
+
+    DEBUG_PRINT("fd(%d) driver '%s' on card '%s' in %s mode", fd, cap.driver, cap.card,
+                IsSinglePlaneVideo(&cap)  ? "splane"
+                : IsMultiPlaneVideo(&cap) ? "mplane"
+                                          : "unknown");
+    gbuffer->fd = fd;
+    gbuffer->type = type;
+    gbuffer->memory = memory;
+    gbuffer->has_dmafd = has_dmafd;
+
+    return true;
+}
+
+bool DequeueBuffer(int fd, v4l2_buffer *buffer) {
+    if (ioctl(fd, VIDIOC_DQBUF, buffer) < 0) {
+        ERROR_PRINT("fd(%d) dequeue buffer: %s", fd, strerror(errno));
+        return false;
+    }
+    return true;
+}
+
+bool QueueBuffer(int fd, v4l2_buffer *buffer) {
+    if (ioctl(fd, VIDIOC_QBUF, buffer) < 0) {
+        ERROR_PRINT("fd(%d) queue buffer(%u): %s\n", fd, buffer->type, strerror(errno));
+        return false;
+    }
+    return true;
+}
+
+bool QueueBuffers(int fd, V4L2BufferGroup *gbuffer) {
+    for (int i = 0; i < gbuffer->num_buffers; i++) {
+        v4l2_buffer *inner = &gbuffer->buffers[i].inner;
+        if (!QueueBuffer(fd, inner)) {
+            return false;
+        }
+    }
+    return true;
+}
+
+bool SubscribeEvent(int fd, uint32_t type) {
+    v4l2_event_subscription sub = {};
+    sub.type = type;
+    if (ioctl(fd, VIDIOC_SUBSCRIBE_EVENT, &sub) < 0) {
+        ERROR_PRINT("fd(%d) does not support VIDIOC_SUBSCRIBE_EVENT(%d)", fd, type);
+        return false;
+    }
+    return true;
+}
+
+bool SetFps(int fd, v4l2_buf_type type, uint32_t fps) {
+    struct v4l2_streamparm streamparms = {};
+    streamparms.type = type;
+    streamparms.parm.output.timeperframe.numerator = 1;
+    streamparms.parm.output.timeperframe.denominator = fps;
+    if (ioctl(fd, VIDIOC_S_PARM, &streamparms) < 0) {
+        ERROR_PRINT("fd(%d) set fps(%d): %s", fd, fps, strerror(errno));
+        return false;
+    }
+    return true;
+}
+
+bool SetFormat(int fd, V4L2BufferGroup *gbuffer, uint32_t width, uint32_t height,
+               uint32_t &pixel_format) {
+    v4l2_format fmt = {};
+    fmt.type = gbuffer->type;
+    ioctl(fd, VIDIOC_G_FMT, &fmt);
+
+    DEBUG_PRINT("fd(%d) original formats: %s(%dx%d)", gbuffer->fd,
+                FourccToString(fmt.fmt.pix_mp.pixelformat).c_str(), fmt.fmt.pix_mp.width,
+                fmt.fmt.pix_mp.height);
+
+    if (width > 0 && height > 0) {
+        fmt.fmt.pix_mp.width = width;
+        fmt.fmt.pix_mp.height = height;
+        fmt.fmt.pix_mp.pixelformat = pixel_format;
+    }
+
+    if (ioctl(fd, VIDIOC_S_FMT, &fmt) < 0) {
+        ERROR_PRINT("fd(%d) set format(%s) : %s", fd,
+                    FourccToString(fmt.fmt.pix_mp.pixelformat).c_str(), strerror(errno));
+        return false;
+    }
+
+    DEBUG_PRINT("fd(%d) latest format: %s(%dx%d)", gbuffer->fd,
+                FourccToString(fmt.fmt.pix_mp.pixelformat).c_str(), fmt.fmt.pix_mp.width,
+                fmt.fmt.pix_mp.height);
+    // use the  return format
+    pixel_format = fmt.fmt.pix_mp.pixelformat;
+    gbuffer->num_planes = fmt.fmt.pix_mp.num_planes;
+
+    if (fmt.fmt.pix_mp.width != width || fmt.fmt.pix_mp.height != height) {
+        ERROR_PRINT("fd(%d) input size (%dx%d) doesn't match driver's output size (%dx%d): %s", fd,
+                    width, height, fmt.fmt.pix_mp.width, fmt.fmt.pix_mp.height, strerror(EINVAL));
+        throw std::runtime_error("the frame size doesn't match");
+    }
+
+    return true;
+}
+
+bool SetCtrl(int fd, uint32_t id, int32_t value) {
+    v4l2_control ctrls = {};
+    ctrls.id = id;
+    ctrls.value = value;
+    if (ioctl(fd, VIDIOC_S_CTRL, &ctrls) < 0) {
+        ERROR_PRINT("fd(%d) set ctrl(%d): %s", fd, id, strerror(errno));
+        return false;
+    }
+    return true;
+}
+
+bool SetExtCtrl(int fd, uint32_t id, int32_t value) {
+    v4l2_ext_controls ctrls = {};
+    v4l2_ext_control ctrl = {};
+
+    /* set ctrls */
+    ctrls.ctrl_class = V4L2_CTRL_CLASS_CODEC;
+    ctrls.controls = &ctrl;
+    ctrls.count = 1;
+
+    /* set ctrl*/
+    ctrl.id = id;
+    ctrl.value = value;
+
+    if (ioctl(fd, VIDIOC_S_EXT_CTRLS, &ctrls) < 0) {
+        ERROR_PRINT("fd(%d) set ext ctrl(%d): %s", fd, id, strerror(errno));
+        return false;
+    }
+    return true;
+}
+
+bool StreamOn(int fd, v4l2_buf_type type) {
+    if (ioctl(fd, VIDIOC_STREAMON, &type) < 0) {
+        ERROR_PRINT("fd(%d) turn on stream: %s", fd, strerror(errno));
+        return false;
+    }
+    return true;
+}
+
+bool StreamOff(int fd, v4l2_buf_type type) {
+    if (ioctl(fd, VIDIOC_STREAMOFF, &type) < 0) {
+        ERROR_PRINT("fd(%d) turn off stream: %s", fd, strerror(errno));
+        return false;
+    }
+    return true;
+}
+
+bool AllocateBuffer(int fd, V4L2BufferGroup *gbuffer, int num_buffers) {
     gbuffer->num_buffers = num_buffers;
     gbuffer->buffers.resize(num_buffers);
 
@@ -322,9 +312,9 @@ bool V4L2Util::AllocateBuffer(int fd, V4L2BufferGroup *gbuffer, int num_buffers)
     return true;
 }
 
-bool V4L2Util::DeallocateBuffer(int fd, V4L2BufferGroup *gbuffer) {
+bool DeallocateBuffer(int fd, V4L2BufferGroup *gbuffer) {
     if (gbuffer->memory == V4L2_MEMORY_MMAP) {
-        V4L2Util::UnMap(gbuffer);
+        UnMap(gbuffer);
     }
 
     v4l2_requestbuffers req = {};
@@ -342,3 +332,5 @@ bool V4L2Util::DeallocateBuffer(int fd, V4L2BufferGroup *gbuffer) {
 
     return true;
 }
+
+} // namespace v4l2_util
