+++ src/rtc/sfu_channel.h
@@ -9,8 +9,6 @@
 #include <api/data_channel_interface.h>
 #include <nlohmann/json.hpp>
 
-#include "common/interface/subject.h"
-#include "common/utils.h"
 #include "ipc/unix_socket_server.h"
 #include "rtc/rtc_channel.h"
